+++ custom_components/alphaess_wallbox/const.py
@@ -3,10 +3,9 @@
 DOMAIN = "alphaess_wallbox"
 
 CONF_SYSTEM_SN = "system_sn"
-CONF_CHARGING_PILE_ID = "charging_pile_id"
+CONF_CHARGER_SN = "charger_sn"
 CONF_SCAN_INTERVAL = "scan_interval"
 
-DEFAULT_CHARGING_PILE_ID = "EV1"
 DEFAULT_SCAN_INTERVAL = 20
 
 SERVICE_SET_CHARGING_MODE = "set_charging_mode"